requirements.txt
pydantic>=2.0
openai>=1.0
chromadb>=0.4
langfuse>=2.0
sentence-transformers>=2.2
python-dotenv
mcp>=0.1.0
# Document OCR dependencies
pdf2image>=1.16.0
Pillow>=10.0.0
anthropic>=0.18.0
